Allow complex members to be updated rather than reset.
When deserializing into an existing object, it is usually the case that any complex members should be preserved. The current behaviour resets any existing members, and only sets the values included in the yaml file - any other properties are reset. The changes here ensure that existing members are preserved, and only properties included in the yaml are changed.

Two tests have been written to clarify the new behaviour.